Refactoring
refactoring

+++ GraphicsEngine/Graphics/Console/ConsoleColorBuffer.cs
@@ -3,16 +3,16 @@
 	public class ConsoleColorBuffer
 		: IConsoleColorBuffer
 	{
-		private readonly byte[,] consoleColorBuffer;
+		private readonly short[,] consoleColorBuffer;
 
 		public ConsoleColorBuffer(int width, int height)
 		{
-			consoleColorBuffer = new byte[width, height];
+			consoleColorBuffer = new short[width, height];
 			Width = width;
 			Height = height;
 		}
 
-		public byte this[int x, int y]
+		public short this[int x, int y]
 		{
 			get { return consoleColorBuffer[x, y]; }
 			set { consoleColorBuffer[x, y] = value; }
@@ -21,9 +21,9 @@
 		public int Width { get; }
 		public int Height { get; }
 
-		public byte[,] GetMultiArrayAsCopy()
+		public short[,] GetMultiArrayAsCopy()
 		{
-			var copy = new byte[Width, Height];
+			var copy = new short[Width, Height];
 			consoleColorBuffer.CopyTo(copy, 0);
 
 			return copy;
